Separated a few key nodes to different files
