+++ dynet/gpu-nodes-matrixmultiply.cu
@@ -1,0 +1,3 @@
+// This is a dummy file that contains the same content as nodes.cc but compiled
+// on CUDA
+#include "nodes-matrixmultiply.cc"
