+++ dynet/nodes.cc
@@ -119,16 +119,8 @@ size_t GaussianNoise::aux_storage_size() const {
   return dim.size() * sizeof(float);
 }
 
-size_t Hinge::aux_storage_size() const {
-  return dim.size() * sizeof(float);
-}
-
 size_t LogSoftmax::aux_storage_size() const {
   return 2 * dim.size() / dim.rows() * sizeof(float);
-}
-
-size_t PickNegLogSoftmax::aux_storage_size() const {
-  return 2 * dim.batch_elems() * sizeof(float) + dim.batch_elems() * sizeof(unsigned int);
 }
 
 // this i need to do something better, but this is a work-around
@@ -261,152 +253,6 @@ void AddVectorToAllColumns::backward_dev_impl(const MyDevice & dev,
   }
 }  
 DYNET_NODE_INST_DEV_IMPL(AddVectorToAllColumns)
-
-// Affine transform uses different implementations for CPU and GPU because this is 
-// much faster than using Eigen's tensor contractions (as of the writing)
-template<class MyDevice>
-void AffineTransform::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
-  DYNET_ASSERT(xs.size() % 2 == 1, "Failed dimension check in AffineTransform::forward");
-  if (xs.size() == 1) {
-    fx.tvec().device(*dev.edevice) = xs[0]->tvec();
-    return;
-  } else {
-    // Add the first matrix
-    size_t b_size = xs[0]->d.size(), fx_size = fx.d.size();
-    if(fx_size == b_size) {
-      fx.tvec().device(*dev.edevice) = xs[0]->tvec();
-    } else {
-#ifdef __CUDACC__
-      Eigen::array<int, 3> bcast; bcast[0] = 1; bcast[1] = fx.d[1]/xs[0]->d[1]; bcast[2] = fx.d.bd/xs[0]->d.bd;
-      fx.tb<2>().device(*dev.edevice) = xs[0]->tb<2>().broadcast(bcast);
-#else
-      DYNET_ARG_CHECK(xs[0]->d.bd == 1, "In AffineTransform, broadcasting over columns with mini-batched inputs is not implemented yet");
-      float *curr_ptr = fx.v, *end_ptr = curr_ptr + fx.d.size(), *in_ptr = xs[0]->v;
-      do {
-        memcpy(curr_ptr, in_ptr, sizeof(float)*b_size);
-        curr_ptr += b_size;
-      } while(curr_ptr != end_ptr);
-#endif
-    }
-
-    // Perform multiplication
-#ifdef __CUDACC__
-    for (unsigned i = 1; i < xs.size(); i += 2)
-      // fx = (acc_sclar)*fx + xs[0] * xs[1]
-      CUDAMatrixMultiply(dev, *xs[i], *xs[i + 1], fx, kSCALAR_ONE);
-#else
-    // Multiply
-    for (unsigned i = 1; i < xs.size(); i += 2) {
-      if(xs[i]->d.bd == 1 && xs[i+1]->d.bd == fx.d.bd) {
-        fx.colbatch_matrix().noalias() += **xs[i] * xs[i+1]->colbatch_matrix();
-      } else {
-        DYNET_ASSERT(xs[i+1]->d.bd == 1 || xs[i+1]->d.bd == xs[i]->d.bd, "Failed dimension check in AffineTransform::forward");
-        for(unsigned b = 0; b < fx.d.bd; ++b) {
-          fx.batch_matrix(b).noalias() += xs[i]->batch_matrix(b) * xs[i+1]->batch_matrix(b);
-        }
-      }
-    }
-#endif
-  }
-}
-
-template<class MyDevice>
-void AffineTransform::backward_dev_impl(const MyDevice & dev,
-                             const vector<const Tensor*>& xs,
-                             const Tensor& fx,
-                             const Tensor& dEdf,
-                             unsigned i,
-                             Tensor& dEdxi) const {
-  DYNET_ASSERT(i < xs.size(), "Failed boundary check in AffineTransform::backward");
-  // Bias term
-  if (i == 0) { // bias term
-    size_t dx_size = dEdxi.d.size(), df_size = dEdf.d.size();
-    if(dx_size == df_size) {
-      dEdxi.tvec().device(*dev.edevice) += dEdf.tvec();
-    } else {
-      DYNET_ARG_CHECK(dEdxi.d.bd == 1, "In AffineTransform, broadcasting over columns with mini-batched inputs is not implemented yet");
-#ifdef __CUDACC__
-      if(dEdxi.d[1] == dEdf.d[1]) {
-        Eigen::array<int, 1> red_axis; red_axis[0] = 2;
-        dEdxi.t<2>().device(*dev.edevice) += dEdf.tb<2>().sum(red_axis);
-      } else {
-        Eigen::array<int, 2> red_axis; red_axis[0] = 1; red_axis[1] = 2;
-        dEdxi.t<1>().device(*dev.edevice) += dEdf.tb<2>().sum(red_axis);
-      }
-#else
-      if(dEdxi.d[1] == dEdf.d[1]) {
-        for(unsigned b = 0; b < dEdf.d.bd; ++b)
-          (*dEdxi).noalias() += dEdf.batch_matrix(b);
-      } else {
-        Tensor mychip(dEdxi.d, dEdf.v, dEdf.device, dEdf.mem_pool);
-        size_t len = dEdf.d.bd * dEdf.d[1];
-        for(unsigned b = 0; b < len; ++b) {
-          (*dEdxi).noalias() += *mychip;
-          mychip.v += dx_size;
-        }
-      }
-#endif
-    }
-
-  // Left argument of matrix multiply
-  } else if (i % 2 == 1) {
-    int max_b = max(dEdf.d.bd, xs[i+1]->d.bd);
-#if __CUDACC__
-    if(dEdxi.d.bd == 1 && (dEdf.d.bd == xs[i+1]->d.bd)) {
-      CUBLAS_CHECK(cublasSgemm(dev.cublas_handle, CUBLAS_OP_N, CUBLAS_OP_T,
-            dEdxi.d.rows(), dEdxi.d.cols(), dEdf.d.cols() * dEdf.d.batch_elems(),
-            kSCALAR_ONE,
-            dEdf.v, dEdf.d.rows(),
-            xs[i+1]->v, xs[i+1]->d.rows(),
-            kSCALAR_ONE, dEdxi.v, dEdxi.d.rows()));
-    } else {
-      for(int b = 0; b < max_b; ++b)
-        CUBLAS_CHECK(cublasSgemm(dev.cublas_handle, CUBLAS_OP_N, CUBLAS_OP_T,
-              dEdxi.d.rows(), dEdxi.d.cols(), dEdf.d.cols(),
-              kSCALAR_ONE,
-              dEdf.batch_ptr(b), dEdf.d.rows(),
-              xs[i+1]->batch_ptr(b), xs[i+1]->d.rows(),
-              kSCALAR_ONE, dEdxi.batch_ptr(b), dEdxi.d.rows()));
-    }
-#else
-    if(dEdxi.d.bd == 1 && (dEdf.d.bd == xs[i+1]->d.bd)) {
-      (*dEdxi).noalias() += dEdf.colbatch_matrix() * xs[i+1]->colbatch_matrix().transpose();
-    } else {
-      for(int b = 0; b < max_b; ++b)
-        dEdxi.batch_matrix(b).noalias() += dEdf.batch_matrix(b) * xs[i+1]->batch_matrix(b).transpose();
-    }
-#endif
-  } else {  // right argument of matrix multiply
-    int max_b = max(xs[i-1]->d.bd, dEdf.d.bd);
-#if __CUDACC__
-    // Do a single multiply if xs[i-1] has one batch
-    if(xs[i-1]->d.bd == 1 && dEdxi.d.bd == dEdf.d.bd) {
-      CUBLAS_CHECK(cublasSgemm(dev.cublas_handle, CUBLAS_OP_T, CUBLAS_OP_N, 
-            dEdxi.d.rows(), dEdxi.d.cols()*dEdxi.d.batch_elems(), xs[i-1]->d.rows(),
-            kSCALAR_ONE,
-            xs[i-1]->v, xs[i-1]->d.rows(),
-            dEdf.v, dEdf.d.rows(),
-            kSCALAR_ONE, dEdxi.v, dEdxi.d.rows()));
-    } else {
-      for(int b = 0; b < max_b; ++b)
-        CUBLAS_CHECK(cublasSgemm(dev.cublas_handle, CUBLAS_OP_T, CUBLAS_OP_N,
-              dEdxi.d.rows(), dEdxi.d.cols(), xs[i-1]->d.rows(),
-              kSCALAR_ONE,
-              xs[i-1]->batch_ptr(b), xs[i-1]->d.rows(),
-              dEdf.batch_ptr(b), dEdf.d.rows(),
-              kSCALAR_ONE, dEdxi.batch_ptr(b), dEdxi.d.rows()));
-    }
-#else
-    if(xs[i-1]->d.bd == 1 && dEdxi.d.bd == dEdf.d.bd) {
-      dEdxi.colbatch_matrix().noalias() += (**xs[i-1]).transpose() * dEdf.colbatch_matrix();
-    } else {
-      for(int b = 0; b < max_b; ++b)
-        dEdxi.batch_matrix(b).noalias() += xs[i-1]->batch_matrix(b).transpose() * dEdf.batch_matrix(b);
-    }
-#endif
-  }
-}
-DYNET_NODE_INST_DEV_IMPL(AffineTransform)
 
 template<class MyDevice>
 void Average::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
@@ -993,71 +839,6 @@ void GaussianNoise::backward_dev_impl(const MyDevice & dev,
 DYNET_NODE_INST_DEV_IMPL(GaussianNoise)
 
 template<class MyDevice>
-void Hinge::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
-  DYNET_ASSERT(xs.size() == 1, "Failed dimension check in Hinge::forward");
-  Tensor eloss(xs[0]->d, static_cast<float*>(aux_mem), fx.device, DeviceMempool::FXS);
-  // TODO: Can we do this on device?
-  if(pelement != nullptr) {
-    DYNET_ARG_CHECK(fx.d.bd == 1, 
-                            "Hinge was passed a single index but the corresponding expression has multiple mini-batch elements (" << fx.d.bd << ")");
-    const real mlystar = margin - TensorTools::access_element(*xs[0], *pelement);
-    eloss.tvec().device(*dev.edevice) = (xs[0]->tvec() + mlystar).cwiseMax(0.f);
-    TensorTools::set_element(eloss, *pelement, 0.f);
-    fx.t<0>().device(*dev.edevice) = eloss.tvec().sum();
-  } else {
-    DYNET_ASSERT(pelements != nullptr, "Hinge::forward has neither pointer to single element nor vector");
-    DYNET_ARG_CHECK(xs[0]->d.bd == pelements->size(),
-                            "The list of indexes passed to Hinge has a length (" << pelements->size() <<
-                            ") that doesn't match the number of mini-batch elements in the corresponding expression (" << xs[0]->d << ")");
-    size_t batch_size = xs[0]->d.batch_size();
-    for(size_t b = 0; b < fx.d.bd; b++) {
-      const real mlystar = margin - TensorTools::access_element(*xs[0], b*batch_size + (*pelements)[b]);
-      eloss.tb<1>().chip<1>(b).device(*dev.edevice) = (xs[0]->tb<1>().chip<1>(b) + mlystar).cwiseMax(0.f);
-      TensorTools::set_element(eloss, b*batch_size + (*pelements)[b], 0.f);
-      fx.tb<0>().chip<0>(b).device(*dev.edevice) = eloss.tb<1>().chip<1>(b).sum();
-    }
-  }
-}
-
-template<class MyDevice>
-void Hinge::backward_dev_impl(const MyDevice & dev,
-                             const vector<const Tensor*>& xs,
-                             const Tensor& fx,
-                             const Tensor& dEdf,
-                             unsigned i,
-                             Tensor& dEdxi) const {
-  DYNET_ASSERT(i == 0, "Failed dimension check in Hinge::backward");
-  if(pelement != nullptr) {
-    if(as_scalar(fx)) { // there was some loss
-      const float d = as_scalar(dEdf);
-      Tensor eloss(xs[0]->d, static_cast<float*>(aux_mem), fx.device, DeviceMempool::FXS);
-      // TODO: The > comparison should not be calculated twice. Keep it in auxiliary memory?
-      dEdxi.tvec().device(*dev.edevice) += (eloss.tvec() > 0.f).cast<float>() * d;
-#if defined(__CUDACC__) && defined(EIGEN_NO_MALLOC)
-      DYNET_RUNTIME_ERR("CUDA memory allocation in hinge");
-#endif
-      dEdxi.tvec().chip<0>(*pelement).device(*dev.edevice) -= (eloss.tvec() > 0.f).template cast<float>().sum() * d;
-    }
-  } else {
-    DYNET_ASSERT(pelements != nullptr, "Hinge::backward has neither pointer to single element nor vector");
-    vector<float> fx_vec = as_vector(fx);
-    vector<float> d_vec = as_vector(dEdf);
-    Tensor eloss(xs[0]->d, static_cast<float*>(aux_mem), fx.device, DeviceMempool::FXS);
-    for(size_t b = 0; b < fx.d.bd; b++) {
-      if(fx_vec[b]) { // there was some loss
-        // TODO: The > comparison should not be calculated twice. Keep it in auxiliary memory?
-        dEdxi.tb<1>().chip<1>(b).device(*dev.edevice) += (eloss.tb<1>().chip<1>(b) > 0.f).cast<float>() * d_vec[b];
-#if defined(__CUDACC__) && defined(EIGEN_NO_MALLOC)
-        DYNET_RUNTIME_ERR("CUDA memory allocation in hinge");
-#endif
-        dEdxi.tb<1>().chip<1>(b).chip<0>((*pelements)[b]).device(*dev.edevice) -= (eloss.tb<1>().chip<1>(b) > 0.f).template cast<float>().sum() * d_vec[b];
-      }
-    }
-  }
-}
-DYNET_NODE_INST_DEV_IMPL(Hinge)
-
-template<class MyDevice>
 void HuberDistance::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
   DYNET_ASSERT(xs.size() == 2, "HuberDistance::forward dimension check failed");
   fx.t<0>().device(*dev.edevice) = (xs[0]->tvec() - xs[1]->tvec()).unaryExpr(FHuberForward(d)).sum();
@@ -1351,95 +1132,6 @@ void MatrixInverse::backward_dev_impl(const MyDevice & dev,
 DYNET_NODE_INST_DEV_IMPL(MatrixInverse)
 
 template<class MyDevice>
-void MatrixMultiply::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
-  DYNET_ASSERT(xs.size() == 2, "Failed dimension check in MatrixMultiply::forward");
-#ifdef __CUDACC__
-  // fx = 0*fx + xs[0] * xs[1]
-  CUDAMatrixMultiply(dev, *xs[0], *xs[1], fx, kSCALAR_ZERO);
-#else
-  DYNET_ASSERT(fx.d.bd == max(xs[0]->d.bd, xs[1]->d.bd), "Failed dimension check in MatrixMultiply::forward");
-  if(xs[0]->d.bd == 1) {
-    // If the left side has one batch, multiply by columns
-    // [x, z, b] = [x, y] * [y, z, b]
-    // -> [x, z*b] = [x, y], [y, z*b]
-    fx.colbatch_matrix().noalias() = **xs[0] * xs[1]->colbatch_matrix();
-  } else {
-    // Otherwise, loop over the batches
-    DYNET_ASSERT(xs[1]->d.bd == 1 || xs[1]->d.bd == xs[0]->d.bd, "Failed dimension check in MatrixMultiply::forward");
-    for(unsigned b = 0; b < xs[0]->d.bd; ++b)
-      fx.batch_matrix(b).noalias() = xs[0]->batch_matrix(b) * xs[1]->batch_matrix(b);
-  }
-#endif
-}
-
-template<class MyDevice>
-void MatrixMultiply::backward_dev_impl(const MyDevice & dev,
-                             const vector<const Tensor*>& xs,
-                             const Tensor& fx,
-                             const Tensor& dEdf,
-                             unsigned i,
-                             Tensor& dEdxi) const {
-  DYNET_ASSERT(i < 2, "Failed dimension check in MatrixMultiply::backward");
-  int max_b = max(xs[0]->d.bd, xs[1]->d.bd);
-#if __CUDACC__
-  if (i == 0) {
-    if(dEdxi.d.bd == 1 && (dEdf.d.bd == xs[1]->d.bd)) {
-      CUBLAS_CHECK(cublasSgemm(dev.cublas_handle, CUBLAS_OP_N, CUBLAS_OP_T,
-            dEdxi.d.rows(), dEdxi.d.cols(), dEdf.d.cols() * dEdf.d.batch_elems(),
-            kSCALAR_ONE,
-            dEdf.v, dEdf.d.rows(),
-            xs[1]->v, xs[1]->d.rows(),
-            kSCALAR_ONE, dEdxi.v, dEdxi.d.rows()));
-    } else {
-      for(int b = 0; b < max_b; ++b)
-        CUBLAS_CHECK(cublasSgemm(dev.cublas_handle, CUBLAS_OP_N, CUBLAS_OP_T,
-              dEdxi.d.rows(), dEdxi.d.cols(), dEdf.d.cols(),
-              kSCALAR_ONE,
-              dEdf.batch_ptr(b), dEdf.d.rows(),
-              xs[1]->batch_ptr(b), xs[1]->d.rows(),
-              kSCALAR_ONE, dEdxi.batch_ptr(b), dEdxi.d.rows()));
-    }
-  } else {
-    // Do a single multiply if xs[0] has one batch
-    if(xs[0]->d.bd == 1) {
-      // dEdxi.colbatch_matrix().noalias() += (**xs[0]).transpose() * dEdf.colbatch_matrix();
-      CUBLAS_CHECK(cublasSgemm(dev.cublas_handle, CUBLAS_OP_T, CUBLAS_OP_N,
-            dEdxi.d.rows(), dEdxi.d.cols()*dEdxi.d.batch_elems(), xs[0]->d.rows(),
-            kSCALAR_ONE,
-            xs[0]->v, xs[0]->d.rows(),
-            dEdf.v, dEdf.d.rows(),
-            kSCALAR_ONE, dEdxi.v, dEdxi.d.rows()));
-    } else {
-      for(int b = 0; b < max_b; ++b)
-        CUBLAS_CHECK(cublasSgemm(dev.cublas_handle, CUBLAS_OP_T, CUBLAS_OP_N,
-              dEdxi.d.rows(), dEdxi.d.cols(), xs[0]->d.rows(),
-              kSCALAR_ONE,
-              xs[0]->batch_ptr(b), xs[0]->d.rows(),
-              dEdf.batch_ptr(b), dEdf.d.rows(),
-              kSCALAR_ONE, dEdxi.batch_ptr(b), dEdxi.d.rows()));
-    }
-  }
-#else
-  if (i == 0) {
-    if(dEdxi.d.bd == 1 && (dEdf.d.bd == xs[1]->d.bd)) {
-      (*dEdxi).noalias() += dEdf.colbatch_matrix() * xs[1]->colbatch_matrix().transpose();
-    } else {
-      for(int b = 0; b < max_b; ++b)
-        dEdxi.batch_matrix(b).noalias() += dEdf.batch_matrix(b) * xs[1]->batch_matrix(b).transpose();
-    }
-  } else {
-    if(xs[0]->d.bd == 1) {
-      dEdxi.colbatch_matrix().noalias() += (**xs[0]).transpose() * dEdf.colbatch_matrix();
-    } else {
-      for(int b = 0; b < max_b; ++b)
-        dEdxi.batch_matrix(b).noalias() += xs[0]->batch_matrix(b).transpose() * dEdf.batch_matrix(b);
-    }
-  }
-#endif
-}
-DYNET_NODE_INST_DEV_IMPL(MatrixMultiply)
-
-template<class MyDevice>
 void Max::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
   Tensor t(fx.d, static_cast<float*>(aux_mem), fx.device, DeviceMempool::FXS);
   t.tvec().device(*dev.edevice) = (xs[0]->tvec() > xs[1]->tvec()).cast<float>();
@@ -1660,71 +1352,6 @@ void PickElement::backward_dev_impl(const MyDevice & dev,
   }
 }
 DYNET_NODE_INST_DEV_IMPL(PickElement)
-
-template<class MyDevice>
-void PickNegLogSoftmax::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
-  if (xs[0]->d.cols() == 1) {
-    Tensor z(Dim({1},fx.d.bd), (float*)aux_mem, fx.device, DeviceMempool::FXS);
-    Tensor m(Dim({1},fx.d.bd), (float*)aux_mem + fx.d.bd, fx.device, DeviceMempool::FXS);
-    unsigned int *ids_dev = (unsigned int*)((float*)aux_mem + 2*fx.d.bd), *ids_host;
-#if __CUDACC__
-    ids_host = (unsigned int*)malloc(fx.d.bd * sizeof(unsigned int));
-#else
-    ids_host = ids_dev;
-#endif
-    if(pval) {
-      *ids_host = *pval;
-    } else {
-      DYNET_ASSERT(pvals, "Neither single nor vector of elements available in PickNegLogSoftmax::forward");
-      DYNET_ARG_CHECK(pvals->size() == fx.d.batch_elems(), 
-                              "In PickNegLogSoftmax::forward, number of elements in the passed-in index vector (" << pvals->size() << ")"
-                              " did not match number of elements in mini-batch elements in expression (of dimension" << fx.d << ")");
-      size_t batch_size = xs[0]->d.batch_size();
-      for(unsigned b = 0; b < fx.d.bd; ++b)
-        ids_host[b] = batch_size * b + (*pvals)[b];
-    }
-#if __CUDACC__
-    CUDA_CHECK(cudaMemcpyAsync(ids_dev, ids_host, fx.d.bd * sizeof(unsigned int), cudaMemcpyHostToDevice));
-    TensorTools::logsumexp_dev(dev, *xs[0], m, z);
-    dynet::gpu::sparse_to_dense_assign(fx.d.bd, ids_dev, xs[0]->v, fx.v);
-    free(ids_host);
-#else
-    TensorTools::logsumexp_dev(dev, *xs[0], m, z);
-    for(unsigned b = 0; b < fx.d.bd; ++b)
-      fx.v[b] = xs[0]->v[ids_dev[b]];
-#endif
-    fx.tvec().device(*dev.edevice) = z.tvec() - fx.tvec();
-  } else {
-    DYNET_RUNTIME_ERR("PickNegLogSoftmax::forward not yet implemented for multiple columns");
-  }
-}
-
-template<class MyDevice>
-void PickNegLogSoftmax::backward_dev_impl(const MyDevice & dev,
-                            const vector<const Tensor*>& xs,
-                            const Tensor& fx,
-                            const Tensor& dEdf,
-                            unsigned i,
-                            Tensor& dEdxi) const {
-  if (xs[0]->d.cols() == 1) {
-    Tensor z(Dim({1},fx.d.batch_elems()), (float*)aux_mem, fx.device, DeviceMempool::FXS);
-    unsigned int *ids_dev = (unsigned int*)((float*)aux_mem + 2*fx.d.bd);
-#if __CUDACC__ 
-    Eigen::array<int, 2> bcast({(int)xs[0]->d[0],1});
-    dEdxi.tb<1>().device(*dev.edevice) += (xs[0]->tb<1>() - z.tb<1>().broadcast(bcast)).exp() * dEdf.tb<1>().broadcast(bcast);
-    dynet::gpu::dense_to_sparse_subtract(fx.d.bd, ids_dev, dEdf.v, dEdxi.v);
-#else
-    // TODO: We want to do broadcasting here too, but it's slow
-    for(unsigned b = 0; b < fx.d.bd; ++b) {
-      dEdxi.tb<1>().chip<1>(b).device(*dev.edevice) += (xs[0]->tb<1>().chip<1>(b) - z.v[b]).exp() * dEdf.v[b];
-      dEdxi.v[ids_dev[b]] -= dEdf.v[b];
-    }
-#endif
-  } else {
-    DYNET_RUNTIME_ERR("PickNegLogSoftmax::backward not yet implemented for multiple columns");
-  }
-}
-DYNET_NODE_INST_DEV_IMPL(PickNegLogSoftmax)
 
 // x_1 is a matrix
 // y = (x_1)[start:end]
